% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.fig.R
\name{save.fig}
\alias{save.fig}
\title{Save plots to tikz figures}
\usage{
save.fig(plt, path, ...)
}
\description{
wrapper over tikzDevice
}
\details{
example:
p <- ggplot(aes(...)) + geom_whatever()
save.fig(p, "figures/plot1.tex")
}
