% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.reg.R
\name{save.reg}
\alias{save.reg}
\title{Save Regression outputs}
\usage{
save.reg(data, path, ...)
}
\description{
wrapper over stargazer-booktabs with some defaults
}
\details{
example:
lm(y ~ x0 + x1 + x2) %>% save.reg("tables/reg1.tex")
}
