% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.tbl.R
\name{save.tbl}
\alias{save.tbl}
\title{Save Tables}
\usage{
save.tbl(data, path, label = NULL, ...)
}
\description{
wrapper over kable that automatically sets labels in output tex
defaults to booktabs
}
\details{
example:
my.df %>% select(col1, col2) %>% save.tbl("tables/table1.tex")
}
